requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.8
MoonrakerPy==0.2.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
requests==2.32.3
urllib3==2.2.2
tqdm==4.67.1
matplotlib==3.10.3
scipy==1.15.3
pylibdmtx